LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-objects/parseAll-expected.txt


FAIL CSSStyleValue.parseAll() with a valid property returns a list with a single CSSStyleValue assert_equals: Result must be a list with one element expected 1 but got 0
FAIL CSSStyleValue.parseAll() is not case sensitive assert_equals: Result must be a list with one element expected 1 but got 0
FAIL CSSStyleValue.parseAll() with a valid list-valued property returns a list with a single CSSStyleValue assert_equals: Result must be a list with three elements expected 3 but got 0
FAIL CSSStyleValue.parseAll() with a valid shorthand property returns a CSSStyleValue 1px 2px 3px 4px cannot be parsed as a margin
FAIL CSSStyleValue.parseAll() with a valid custom property returns a list with a single CSSStyleValue Property String is not a valid CSS property.
